Refactored multiple default selects into a component to cut down on repeated code.
