+++ Source/Applications/SystemCenter/SystemCenter.csproj
@@ -434,6 +434,7 @@
       <SubType>Designer</SubType>
     </None>
     <Content Include="wwwroot\Scripts\TSX\SystemCenter\Location\LocationImages.tsx" />
+    <Content Include="wwwroot\Scripts\TSX\SystemCenter\Asset\AssetSelect.tsx" />
     <TypeScriptCompile Include="wwwroot\Scripts\TSX\SystemCenter\hooks.ts" />
     <TypeScriptCompile Include="wwwroot\Scripts\TSX\SystemCenter\RemoteXDA\RemoteAssetForm.tsx" />
     <TypeScriptCompile Include="wwwroot\Scripts\TSX\SystemCenter\RemoteXDA\RemoteMeterTab.tsx" />
