+++ Source/Applications/SystemCenter/wwwroot/Scripts/TSX/SystemCenter/Asset/AssetSelect.tsx
@@ -1,0 +1,127 @@
+﻿//******************************************************************************************************
+//  AssetSelect.tsx - Gbtc
+//
+//  Copyright © 2022, Grid Protection Alliance.  All Rights Reserved.
+//
+//  Licensed to the Grid Protection Alliance (GPA) under one or more contributor license agreements. See
+//  the NOTICE file distributed with this work for additional information regarding copyright ownership.
+//  The GPA licenses this file to you under the MIT License (MIT), the "License"; you may not use this
+//  file except in compliance with the License. You may obtain a copy of the License at:
+//
+//      http://opensource.org/licenses/MIT
+//
+//  Unless agreed to in writing, the subject software distributed under the License is distributed on an
+//  "AS-IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. Refer to the
+//  License for the specific language governing permissions and limitations.
+//
+//  Code Modification History:
+//  ----------------------------------------------------------------------------------------------------
+//  11/07/2022 - G. Santos
+//       Generated original version of source code.
+//
+//******************************************************************************************************
+
+
+import * as React from 'react';
+import * as _ from 'lodash';
+import { ByAssetSlice, AssetTypeSlice } from '../Store/Store';
+import { SystemCenter } from '@gpa-gemstone/application-typings';
+import { Search } from '@gpa-gemstone/react-interactive';
+import { DefaultSelects } from '@gpa-gemstone/common-pages';
+import { useAppDispatch, useAppSelector } from '../hooks';
+
+declare var homePath: string;
+
+interface IProps {
+    OnCloseFunction: (selectedArrays: Array<SystemCenter.Types.DetailedAsset>, confirmed: boolean) => void,
+    SelectedAssets: Array<SystemCenter.Types.DetailedAsset>,
+    ShowModal: boolean,
+    Type?: 'single' | 'multiple',
+    SessionStorageID?: string, //TODO: Add this functionality to save filters in session storage (gemstone addition)
+    Title?: string
+}
+
+export default function AssetSelect(props: IProps) {
+    const assetType = useAppSelector(AssetTypeSlice.Data);
+    const assetTypeStatus = useAppSelector(AssetTypeSlice.Status);
+    const dispatch = useAppDispatch();
+
+    React.useEffect(() => {
+        if (assetTypeStatus == 'changed' || assetTypeStatus == 'unintiated')
+            dispatch(AssetTypeSlice.Fetch());
+    }, [assetTypeStatus]);
+
+    function getEnum(setOptions, field) {
+        if (field.key == 'AssetType' && field.type == 'enum') {
+            setOptions(assetType.map((t) => ({ Value: t.Name, Label: t.Name })))
+            return () => { }
+        }
+
+        if (field.type != 'enum' || field.enum == undefined || field.enum.length != 1)
+            return () => { };
+
+        let handle = $.ajax({
+            type: "GET",
+            url: `${homePath}api/ValueList/Group/${field.enum[0].Value}`,
+            contentType: "application/json; charset=utf-8",
+            dataType: 'json',
+            cache: true,
+            async: true
+        });
+
+        handle.done(d => setOptions(d.map(item => ({ Value: item.Value.toString(), Label: item.Text }))))
+        return () => {
+            if (handle != null && handle.abort == null) handle.abort();
+        }
+    }
+
+    function getAdditionalAssetFields(setFields) {
+        let handle = $.ajax({
+            type: "GET",
+            url: `${homePath}api/SystemCenter/AdditionalField/ParentTable/Asset/FieldName/0`,
+            contentType: "application/json; charset=utf-8",
+            cache: false,
+            async: true
+        });
+
+        function ConvertType(type: string) {
+            if (type == 'string' || type == 'integer' || type == 'number' || type == 'datetime' || type == 'boolean')
+                return { type: type }
+            return {
+                type: 'enum', enum: [{ Label: type, Value: type }]
+            }
+        }
+
+        handle.done((d: Array<SystemCenter.Types.AdditionalField>) => {
+
+            let ordered = _.orderBy(d.filter(item => item.Searchable).map(item => (
+                { label: `[AF${item.ExternalDB != undefined ? " " + item.ExternalDB : ''}] ${item.FieldName}`, key: item.FieldName, ...ConvertType(item.Type), isPivotField: true } as Search.IField<SystemCenter.Types.DetailedAsset>
+            )), ['label'], ["asc"]);
+            setFields(ordered);
+        });
+        return () => {
+            if (handle != null && handle.abort == null) handle.abort();
+        };
+    }
+
+    return (
+        <DefaultSelects.Asset
+            Slice={ByAssetSlice}
+            Selection={props.SelectedAssets}
+            OnClose={props.OnCloseFunction}
+            Show={props.ShowModal}
+            Type={props.Type === undefined ? 'multiple' : props.Type}
+            Columns={[
+                { key: 'AssetKey', field: 'AssetKey', label: 'Key', headerStyle: { width: 'auto' }, rowStyle: { width: 'auto' } },
+                { key: 'AssetName', field: 'AssetName', label: 'Name', headerStyle: { width: 'auto' }, rowStyle: { width: 'auto' } },
+                { key: 'AssetType', field: 'AssetType', label: 'Asset Type', headerStyle: { width: 'auto' }, rowStyle: { width: 'auto' } },
+                { key: 'VoltageKV', field: 'VoltageKV', label: 'Voltage (kV)', headerStyle: { width: 'auto' }, rowStyle: { width: 'auto' } },
+                { key: 'Meters', field: 'Meters', label: 'Meters', headerStyle: { width: 'auto' }, rowStyle: { width: 'auto' } },
+                { key: 'Locations', field: 'Locations', label: 'Substations', headerStyle: { width: 'auto' }, rowStyle: { width: 'auto' } },
+                { key: 'Scroll', label: '', headerStyle: { width: 17, padding: 0 }, rowStyle: { width: 0, padding: 0 } },
+            ]}
+            Title={props.Title === undefined ? "Select Assets" : props.Title}
+            GetEnum={getEnum}
+            GetAddlFields={getAdditionalAssetFields} />
+    )
+}
